Downgrade deps to work around crash due to incorrect usage of Xlib
AFter the last commit the game crashes on startup on my laptop,
with the error:
[xcb] Unknown sequence number while processing queue
[xcb] Most likely this is a multi-threaded client and XInitThreads has not been called
[xcb] Aborting, sorry about that.
space_tennis: ../../src/xcb_io.c:260: poll_for_event: Assertion `!xcb_xlib_threads_sequence_lost' failed.
Aborted (core dumped)

The only difference between piston_window 0.127 and 0.128 was bumping
pistoncore-glutin_window from 0.70 to 0.71.
And the biggest change in that version was upgrading glutin from 0.26.0
to 0.30.6.
glutin 0.27 stopped using winit, which does call XInitThreads(),
so it seems like the glutin devs forgot to start calling it again,
because glutin was calling it before it switched to winin:
https://github.com/rust-windowing/glutin/commit/28a30dcb84c84d0dae62af68ee0dbde87f2c654e
But pistoncore-glutin-window 0.71 also added winit as dependency,
so I'm not sure which library should be calling it.

Since piston_window now has pistoncore-glutin_window as default feature,
stop depending on that directly.