Switch to speedy2d game engine
Because it supports wasm, which piuton doesn't.
But speedy2d doesn't support android or ios, which piston does,
so leve the code for piston in place in a way that can be enabled
with minimal changes to engine/Cargo.toml

Because speedy2d lacks the update event, we have to trigger custom
events from a different thread.
This means there might be some more jitter in handling input,
but it's the best we can do.

Speedy2d doesn't have a method for drawing ellipses,
but only has one for circles.
This means we can't stretch it when window is resized to a
non-square aspect ratio. So here we have to compromise a little.